refactor: rename model classes and update repository methods
