+++ chat/src/main/kotlin/net/thechance/chat/api/dto/ChatResponse.kt
@@ -1,15 +1,15 @@
-package net.thechance.chat.service.model
+package net.thechance.chat.api.dto
 
 import net.thechance.chat.entity.Chat
 import java.util.UUID
 
-class ChatModel(
+class ChatResponse(
     val id: UUID,
     val name: String,
     val imageUrl: String?,
 )
 
-fun Chat.toModel(requesterId: UUID) = ChatModel(
+fun Chat.toResponse(requesterId: UUID) = ChatResponse(
     id = id,
     name = users.firstOrNull { it.id != requesterId }
         ?.let { "${it.firstName} ${it.lastName}" }
