+++ chat/src/main/kotlin/net/thechance/chat/api/dto/ContactResponse.kt
@@ -1,5 +1,6 @@
 package net.thechance.chat.api.dto
 
+import net.thechance.chat.service.model.ContactModel
 import java.util.UUID
 
 data class ContactResponse(
@@ -7,7 +8,17 @@ data class ContactResponse(
     val firstName: String,
     val lastName: String,
     val phoneNumber: String,
-    val isMenaUser: Boolean,
     val userId: UUID?,
     val imageUrl: String?
 )
+
+fun ContactModel.toResponse() : ContactResponse {
+    return ContactResponse(
+        id = this.id,
+        firstName = this.firstName,
+        lastName = this.lastName,
+        phoneNumber = this.phoneNumber,
+        userId = this.userId,
+        imageUrl = this.imageUrl
+    )
+}
