+++ chat/src/main/kotlin/net/thechance/chat/api/dto/PagedResponse.kt
@@ -1,7 +1,6 @@
 package net.thechance.chat.api.dto
 
 import net.thechance.chat.service.model.ContactModel
-import net.thechance.chat.service.model.toResponse
 import org.springframework.data.domain.Page
 
 
@@ -16,7 +15,7 @@ data class PagedResponse<T>(
 fun Page<ContactModel>.toResponse(): PagedResponse<ContactResponse>{
     return PagedResponse(
         data = this.content.map { it.toResponse() },
-        pageNumber = this.number +1,
+        pageNumber = this.number + 1,
         pageSize = this.size,
         totalItems = this.totalElements,
         totalPages = this.totalPages
