+++ chat/src/main/kotlin/net/thechance/chat/entity/ContactUser.kt
@@ -3,7 +3,6 @@ package net.thechance.chat.entity
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.Id
-import jakarta.persistence.ManyToMany
 import jakarta.persistence.Table
 import java.util.UUID
 
@@ -25,7 +24,4 @@ data class ContactUser(
 
     @Column(nullable = true)
     val imageUrl: String? = null,
-
-    @ManyToMany(mappedBy = "users")
-    val chats: Set<Chat> = emptySet(),
 )
