+++ chat/src/main/kotlin/net/thechance/chat/repository/ChatRepository.kt
@@ -6,6 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository
 import java.util.UUID
 
 interface ChatRepository : JpaRepository<Chat, UUID> {
-    fun findByIdIs(id: UUID): Chat?
     fun findByUsers(users: Set<ContactUser>): Chat?
 }
