+++ chat/src/main/kotlin/net/thechance/chat/service/model/ContactModel.kt
@@ -1,6 +1,5 @@
 package net.thechance.chat.service.model
 
-import net.thechance.chat.api.dto.ContactResponse
 import java.util.UUID
 
 data class ContactModel(
@@ -8,19 +7,6 @@ data class ContactModel(
     val firstName: String,
     val lastName: String,
     val phoneNumber: String,
-    val isMenaUser: Boolean,
     val userId: UUID?,
     val imageUrl: String?
 )
-
-fun ContactModel.toResponse() : ContactResponse{
-    return ContactResponse(
-        id = this.id,
-        firstName = this.firstName,
-        lastName = this.lastName,
-        phoneNumber = this.phoneNumber,
-        isMenaUser = this.isMenaUser,
-        userId = this.userId,
-        imageUrl = this.imageUrl
-    )
-}
